requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
autopep8==2.3.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
et-xmlfile==1.1.0
FLAML==2.1.2
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.51.0
h11==0.14.0
hedtools @ git+https://github.com/hed-standard/hed-python/@c61c7afa52e9a3930b2f706cd0d7dce9574954e2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
inflect==7.2.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jdcal==1.4.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.17
numpy==1.26.4
openai==1.30.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pkginfo==1.11.1
portalocker==2.8.2
pyautogen==0.2.29
pycodestyle==2.12.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
readme_renderer==43.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
twine==5.1.0
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wordcloud==1.9.3
zipp==3.19.2
